Add files via upload
Changes include:
Full-screen video that is responsive.  Still need to add media queries for mobile.  Use poster-video for image of video - do not play video on mobile.
Added overlay color and overlay name and title, with About Me button that doesn't take the user anywhere yet.  Fully responsive!
Logo rotates with animation in navbar.  Still need to adjust the way the navbar elements on the right look with css.
Not positive that the kickstarter project section is how I want it to look.  Might revisit later.  For now, there is an image and some text.  All responsive.
The second section is added but no real content for it yet.  Just some Lorem ipsum text.